[Cashtab] [chronik utxo handling p12] etoken send txs with chronik formatted utxos
Summary:
T2447

Depends on D11979

Support for sending eToken txs with new chronik utxo formatting.

Note: there is a bug now on the SendToken screen where the UI unlocks before the balance updates. This will be fixed in the next diff on this stack.

Test Plan:
`npm start`
Send several eToken txs
Check them in the block explorer for sanity (correct fee, change, etc)

Note that the UI of SendToken will sometimes unlock before the balance has updated. This will be corrected in the next diff on this stack.

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D11985