[avalanche] Fix an abc_p2p_proof_inventory flakiness
Summary:
When checking that all peers received a proof inventory message, the last_message attribute is used. This is problematic because this will fail as soon as another inventory is received by any of the peer while the others didn't all have the proof yet. This scenario is especially likely to happen because the generated proof mines a block that needs to be relayed.

This diff remembers the last received proofid so that if another inventory message is received it won't break the test.

Test Plan:
  ./test/functional/test_runner.py abc_p2p_proof_inventory

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D12029