[Cashtab] Move token burn UI lock to same as other token tx method
Summary:
T2670

Cashtab locks the UI after a tx is broadcast until the utxo set has been updated. There is a bug on the burn token tx feature where the UI is unlocked manually.

This should be removed so the UI unlocks in the same way as after any other tx, by the `useEffect` loop that waits for balance to update.

Test Plan:
`npm start`
Burn some eToken
Note that the UI locks and restores after utxo set is updated

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D12047