[avalanche] Fix a abc_feature_proof_cleanup potential failure
Summary: The test expires the proof dangling timeout, then look at the log while elapsing the scheduler interval and expects the proofs to be removed. It can happen (https://build.bitcoinabc.org/viewLog.html?buildTypeId=BitcoinABC_BitcoinAbcStaging&buildId=440597&guest=1) that the scheduler runs after the timeout is elapsed but before the log is getting asserted, causing the test to fail. Elapsing the timeout after the log is read will prevent this.

Test Plan:
  ./test/functional/test_runner.py abc_feature_proof_cleanup

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D12082