[avalanche] Clarify the abc_p2p_compactproofs test
Summary: A wild CI failure happened on this test (https://build.bitcoinabc.org/viewLog.html?buildTypeId=BitcoinABC_Master_BitcoinAbcMasterWOWallet&buildId=lastFinished&guest=1#testNameId-799256654276768243) when a message took just over 2s to appear in the logs, causing the assert_debug_log to fail on timeout. This is an opportunity to remove the log lookup entirely and clarify the intent of the test alltogether.

Test Plan:
  ./test/functional/test_runner.py abc_p2p_compactproof

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D12083