[Cashtab] Patched nav menu being cut off on some mobile views
Summary: Related to T2703.  On some mobile views the nav menu was being cut off on the right edge. NavMenu styling has been changed to set the div right: 0 when in mobile view.

Test Plan:
navigate to https://cashtab-dev-test.netlify.app/#/wallet on a mobile device.
open the nav menu and observe that it is fully visible and on the right most edge of the page

Reviewers: emack, bytesofman, #bitcoin_abc

Reviewed By: emack, bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D12100