Prepare for using debian 11 for the regular builds
Summary:
This diff updates the setup files to work with Debian 11. This is a step
toward migrating our infra to Debian 11.

The files are deliberately duplicated so it doesn't break the CI, and
the buster variant will be removed after the infra has been updated to
use bullseye.

The build instructions are updated to work with Debian 11 accordingly.

Test Plan:
Run the install script on Debian 11.
Read the docs.

Reviewers: #bitcoin_abc, PiRK, sdulfari

Reviewed By: #bitcoin_abc, PiRK, sdulfari

Subscribers: sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D12104