[Cashtab] [chronik tx history p11] Update incoming tx notifications parsing to use new function
Summary:
T2447

Depends on D12130

The parseChronikTx function was originally used only to parse incoming websocket txs for incoming tx notifications. This used a legacy way of getting token info.

Modify this implementation to use the now existing cashtabCache object.

Test Plan:
Review changes to unit tests and `npm test`
`npm start`
Receive an etoken tx of an etoken already in your cache, observe notification is correct
Receive an etoken tx of an etoken not in your cache, observe notification is correct and cashtabCache is updated in local storage

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D12131