Increase max ipv4 seeds per ASN
Summary:
With avalanche it is more common for nodes to be run in a datacenter, making
the current ipv4 per asn limit too small. At current network size and using
https://bitnodes.io/nodes/all/asns/ as a guide, a limit of 6 is sufficient to
record nearly the entire network as seeds assuming it follows the distribution
in that link. Larger limits do not increase the number of seeds outputted so
there is no reason to go higher at this point in time.

Depends on D12141

Test Plan:
  python3 makeseeds.py < seeds_main.txt > nodes_main.txt
  wc -l nodes_main.txt
  Before patch: 18
  After patch: 48

  python3 makeseeds.py < seeds_test.txt > nodes_test.txt
  wc -l nodes_test.txt
  Before patch: 10
  After patch: 14

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D12157