[Cashtab] [chronik tx history p21] Better token burn parsing
Summary:
T2447

Depends on D12158

For token transactions with high decimal count, chronik will (sometimes?) include the `slpBurn` field, even if the burned amount is 0...so it's not actually a burn.

This specifies that the existence of an `slpBurn` field is not enough to imply a token burn. Amount must be zero. This corrects an issue where non-burn txs were being parsed incorrectly (wrong amount, wrong tx type)

Test Plan:
Review new unit test and `npm test`
Receive an etoken tx with 9 decimal places
Verify that it is rendered in tx history as a received etoken tx with correct quantity

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D12159