[Cashtab] Patch issue with burn all of a token locking UI
Summary:
T2716

In the case of a burn token tx that burns all of a token balance, the `useEffect` loop that unlocks the UI was not being hit. Add a simple conditional which catches this case.

Test Plan:
`npm start`
Burn all of a token and note that UI unlocks at same time app redirects to home
Burn some of a token and observe UI unlock functions as expected
Send some of a token and observe UI unlock functions as expected

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D12161