Fix localhost typo in addresses directed at addnode rpc
Summary:
addnode has no input validation so will accept garbage inputs but fail later
in the network stack. This case just happens to work since the network code
interprets 127.0.01 as 127.0.0.1. But it's obviously wrong and makes the logs
look incorrect so it is worth fixing. There is no change in behavior.

Bitcoin Core is aware that addnode needs improvements. Since the impact of lack
of validation is low, we can either wait until a backport is available or fix
the root cause as resourcing allows.

Test Plan:
  ./test/functional/test_runner.py abc_p2p_*

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D12167