[avalanche] Avoid connecting via addavalanchenode in tests
Summary:
This splits the get_ava_p2p_interface in 2 flavors: with or without handshake. This allows for using the natural peer discovery for connecting the avalanche peers in tests rather than requiring to manual connection via addavalanchenode. The tests that can benefit from a noban whitelist to speed up the tests gets updated as well.

Depends on D12176.

Test Plan:
  ninja check-functional-extended

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Subscribers: sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D12177