[Cashtab][Pt 2/2] Parse Airdrops in TxHistory - using cashtabCache to get airdrop token name
Summary:
Added cashtabCache to context exports in useWallet.js, passed cashtabCache through Home.js to Tx.js. Added logic in Tx.js to look for the index of data.parsed.airdropTokenId in cashtabCache.tokenInfoById, if it is found, the tokenName is added to the Tx.

Depends on D12195

Test Plan:
cd web/cashtab && npm start
if there are no sent or received airdrop txs in recent tx history, make them
observe that the token name is now visible under the token icon

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D12196