[Cashtab] [extension address passing p4] Create GetAddress.tsx stub with stories
Summary:
T2550

Depends on D12316

Create a stub component 'GetAddress'. This will be built on later in the stack to fetch and display the extension's active wallet address.

Test Plan:
```
cd web/cashtab-components
npm run storybook
```
Navigate to the `GetAddress` component
Open the dev console
Click the button and note `handleGetAddress` appears in the dev console

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D12317