[Cashtab] [extension address passing p6] Extension runs an address fetch function when requested to do so
Summary:
T2550

Depends on D12318

Update GetAddress.tsx so that handleGetAddress sends an address request to the browser extension.
Updage background.js in the browser extension so that it recognizes this request and runs this function.

Test Plan:
In one terminal,
```
cd web/cashtab-components
npm run storybook
```

In an other terminal,
```
cd web/cashtab
npm run extension
```

Navigate to `brave://extension` and click 'update'
Click 'inspect views in background page'
This will open a small window with a dev console

In your local storybook, click on the 'Get Address' button in the GetAddress component
Note that the dev console for `background.js` reads "User is requesting extension address" and "fetchAddress called in background.js'

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D12319