[Cashtab Extension] fixed roaming nav menu when extension popped out
Summary: Related to T2746 & T2743. Fixed issue with the new hamburger menu roaming in the extension when it is popped out.

Test Plan:
cd web/cashtab
npm run extension
load new build
open the extension and click the pop out button from the top right corner
observe that the menu is in the correct position

Reviewers: bytesofman, emack, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D12354