[Cashtab] Token balance should show full decimals
Summary: Related to T2722. In order to show the trailing zeros on tokenBalance, a new function has been added. As the optionalLocale paramater was not being used in any of the instances of formatBalance related to eTokens, the new function does not accept optionalLocale as a param. If this is a good way forward I will add the unit tests for the new function.

Test Plan:
cd web/cashtab && npm start
navigate to eTokens page
observe that for each token balance, the decimals, including trailing zeros, are kept as part of the balance.
navigate to sendToken page
observe that for the token balance, the decimals, including trailing zeros, are kept as part of the balance.

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D12392