[Cashtab][Pt 2/4] Move message signing and verification to new screen - Remove Verification & Signing from Send.js
Summary: Related to T2734, depends on D12426. Message signing and verification collapses should be on their on screen. Part 2 of this diff is the removal of these collapses from Send.js

Test Plan:
cd web/cashtab && npm start
navigate to the Send screen
observe that the message verification and signing collapses are no longer present

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D12428