Fix rpc_blockchain flakiness
Summary: The test use an assertion that a key doesn't exist: `assert 'fee' not in block['tx'][1]` when the verbosity level is 1. This expected to work because with this verbosity level, the output is an array of txid and not a dict, so the key cannot be found. Except that `fee` is a valid hexadecimal value, and nothing prevents the txid from having this string.

Test Plan:
  for i in {1..100}; do ./test/functional/test_runner.py rpc_blockchain || break; done

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: PiRK

Differential Revision: https://reviews.bitcoinabc.org/D12445