[Cashtab] Visual navigation aid added for hamburger menu options
Summary: Related to T2774. Added visual navigation aid when user is on one of the screens accessed through the hamburger menu.

Test Plan:
cd web/cashtab && npm start
navigate to the Airdrop screen
observe that the Airdrop icon is present on the bottom right hand side of the screen
navigate to the SignVerifyMsg screen
observe that the SignVerifyMsg icon is present on the bottom right hand side of the screen
navigate to the Settings screen
observe that the Settings icon is present on the bottom right hand side of the screen

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D12473