[Cashtab] [p1 local msg verify] Verify function is not async
Summary:
T2730

Stacked diff to deprecate `bchObj.BitcoinCash.verifyMessage`

This diff removes the `async` from the existing function, as `bchObj.BitcoinCash.verifyMessage` is not an async function. This is a bug fix as preparation for dropping in a replacement function.

Test Plan:
`npm start`
Verify some messages, note no change in existing functionality

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D12458