[Cashtab] [p8 local msg verify] Deprecate legacy verification function
Summary:
T2730

Depends on D12464

Remove the legacy approach at signature verification

There is an unrelated bug where the message signing function is using `BCH` instead of `bchObj`. This would be fixed in another diff, except that we would have a build error in this diff if we left `bchObj` unused. So, I've patched it here.

Test Plan:
`npm start`
Sign and verify a message in Cashtab
Confirm cashtab.com also verifies the signature

Sign a message in ElectrumABC using an ecash signature
Verify locally with this diff
Verify that cashtab.com calls it false

Reviewers: emack, #bitcoin_abc

Reviewed By: emack, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D12465