[avalanche] make the mocktime increase monotonically
Summary:
The logic was wrong in this test, and the mocktime was set back in time. `MAX_ADDR_SEND_DELAY` is larger than `GETAVAADDR_INTERVAL`, so it is not necessary to change the mocktime after already moving the time far enough after the first `GETAVAADDR`.

I added an assertion to make sure we will get a meaningful error if `GETAVAADDR_INTERVAL` ever becomes larger than `MAX_ADDR_SEND_DELAY`.

Test Plan: `ninja check-functional`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D12523