[avalanche] decrease the number of GETAVAADDR spamming iterations and  bump the timeout
Summary:
This test sometimes fails n CI because the expected log message is seen after the default `assert_debug_log` 2 seconds timeout. Bump the timeout to make the failure less likely to happen.
Also reduce the number of iterations to 2 to speedup the test. The choice of 10 iterations was arbitrary, and does not prove anything.

Test Plan: `ninja check-functional`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D12524